feat: add proper authorization and error handling

+++ server/api/queue/[id].delete.ts
@@ -1,6 +1,9 @@
 import prisma from '~/lib/prisma'
+import authorize from '~/server/utils/auth'
 
 export default defineEventHandler(async (event) => {
+  await authorize(event, Role.MANAGER)
+
   const parsedQueueId = ZQueueId.safeParse(getRouterParam(event, 'id'))
   if (parsedQueueId.error) throw parsedQueueId.error
   return await prisma.queue.delete({ where: { id: parsedQueueId.data } })
